Split dom package
