+++ hyperscript/object.go
