+++ hyperscript/text.go
